FIX & FEAT: faces parsed properly, more precise triangle draw and random color attribution

+++ canvas.hpp
@@ -27,7 +27,7 @@ class Canvas{
 		bool IsInTriangle(Vec2 vertex_1, Vec2 vertex_2, Vec2 vertex_3, Vec2 position);
 		bool AABB_Collision(int min_x, int max_x, int min_y, int max_y);
 
-		int LeftRightVector(Vec2 A, Vec2 B, Vec2 P);
+		float LeftRightVector(Vec2 A, Vec2 B, Vec2 P);
     Vec2 ScreenSpacePerspectiveProjection(Vec3 vertex);
 		void DrawTriangle(Triangle* triangle);
 };
